feat: Complete project architecture overhaul
