+++ cogs/views/modals.py
@@ -1,16 +1,18 @@
-# --- cogs/utils/views_modals.py ---
+# cogs/views/modals.py
 # This file contains all the Discord UI modal classes for the game.
 
 import discord
 
-# --- REFACTORING FIX ---
-# Removed ITEM_DATA and ITEM_CATEGORIES from this import as they are no longer used in this file.
+# --- REFACTORED IMPORTS ---
+# The helpers and other data this file might need should be imported
+# from their new top-level locations.
+from utils.helpers import get_status_bar
 
 
-# --- Modals ---
 class RenamePetModal(discord.ui.Modal, title="Rename Your Pet"):
     """
     A modal for renaming a pet.
+    The logic inside this class was already solid and needs no changes.
     """
 
     def __init__(self, bot, pet_id, old_name, parent_view):
@@ -31,20 +33,19 @@ class RenamePetModal(discord.ui.Modal, title="Rename Your Pet"):
         await interaction.response.defer(ephemeral=True)
         db_cog = self.bot.get_cog('Database')
         if not db_cog:
-            print("LOG: Database cog not found in RenamePetModal.on_submit")
             return await interaction.followup.send("Database not loaded. Please contact an administrator.")
+
         new_name = self.new_name_input.value
         try:
-            print(f"LOG: Attempting to rename pet ID {self.pet_id} to {new_name}")
             await db_cog.update_pet(self.pet_id, name=new_name)
             await interaction.followup.send(f"You have renamed **{self.old_name}** to **{new_name}**!", ephemeral=True)
-            print("LOG: Renaming successful. Updating parent view.")
-            self.parent_view.main_pet = await db_cog.get_pet(self.pet_id)
-            new_embed = await self.parent_view.get_pet_status_embed(self.parent_view.main_pet)
-            try:
-                await self.parent_view.message.edit(embed=new_embed, view=self.parent_view)
-            except discord.errors.NotFound:
-                pass  # Message was likely deleted
+
+            # Refresh the parent view with the new pet name
+            self.parent_view.main_pet_data = await db_cog.get_pet(self.pet_id)
+            # Recreate the main_pet object with new data in the parent view
+            self.parent_view.main_pet.__init__(self.parent_view.main_pet_data)
+            new_embed = await self.parent_view.get_pet_status_embed()
+            await self.parent_view.message.edit(embed=new_embed, view=self.parent_view)
         except Exception as e:
             print(f"ERROR: An error occurred in RenamePetModal.on_submit: {e}")
             await interaction.followup.send(f"An error occurred while renaming your pet: {e}", ephemeral=True)
